% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibration_questions}
\alias{calibration_questions}
\title{Calibration questions}
\format{
A data frame with 27 rows and 3 variables:
\describe{
\item{question}{text of the calibration question}
\item{answer}{answer text to the calibration question}
\item{calibration_id}{unique identifier for the calibration question}
}
}
\source{
Common trivia questions drawn from a variety of open source web resources.
}
\usage{
calibration_questions
}
\description{
A dataset of reference trivia questions for calibrating SMEs.
}
\keyword{datasets}
