% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_readability}
\alias{check_readability}
\title{Check the readability of scenario text}
\usage{
check_readability(x)
}
\arguments{
\item{x}{A \code{tidyrisk_question_set} object}
}
\value{
A dataframe of the scenario id, domain id, and the Flesch-Kincaid readability score for the scenario text.
}
\description{
Calculate the Flesch-Kincaid score for each scenario and return that score
along with the scenario ID and domain as a tidy dataframe.
}
\examples{
\dontrun{
questions <- read_questions()
check_readability(questions)
}
}
