% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_answers}
\alias{clean_answers}
\title{Clean the answers data frames}
\usage{
clean_answers(capability_answers, scenario_answers)
}
\arguments{
\item{capability_answers}{Capability answers.}

\item{scenario_answers}{Scenario answers.}
}
\value{
A list.
}
\description{
Make the following assumptions/modifications
  - minimum capacity is 5% (we've thought about it - 90% CI)
  - maximum capacity is 95% (we're just about the best - 90% CI)
  - minimum loss is 1000 dollars (both low and high)
  - scale all impact into thousands of dollars (make normal
      decomposition easier, and is in line of the scale of
      a strategic analysis)
  - set a minimum frequency of once per 10 years (0.1)
}
\examples{
NULL
}
