% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collector.R
\name{collect_and_run}
\alias{collect_and_run}
\title{Collect arg and call, and call original function}
\usage{
collect_and_run()
}
\value{
The same output as the original function, in addition to side effects.
}
\description{
This is not meant to be called by the end user. \code{set_collector()} will
place those calls
}
