% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collector-package.R
\docType{package}
\name{collector-package}
\alias{collector}
\alias{collector-package}
\title{collector: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cynkra/collector}
  \item Report bugs at \url{https://github.com/cynkra/collector/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
