% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collector.R
\docType{package}
\name{collector}
\alias{collector}
\alias{collector-package}
\title{\code{collector} package}
\description{
Quantified Information Risk Assessment Data Collection
}
\details{
See the online documentation located at
\href{https://evaluator.tidyrisk.org/}{https://evaluator.tidyrisk.org/}
}
