% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{combine_capability_parameters}
\alias{combine_capability_parameters}
\title{Combine multiple SME distributions into a single unified view}
\usage{
combine_capability_parameters(capability_parameters)
}
\arguments{
\item{capability_parameters}{Fitted individual parameters for capabilities.}
}
\value{
A dataframe.
}
\description{
Given a dataframe with multiple SME fitted distributions for a single
  capability, apply weighting for opinion pooling, and contruct a final
  combined distribution for each OpenFAIR scenario parameter.
}
\examples{
NULL
}
