% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{combine_lognorm}
\alias{combine_lognorm}
\title{Weight a set of lognormal parameters into a single distribution}
\usage{
combine_lognorm(dat)
}
\arguments{
\item{dat}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Weight a set of lognormal parameters into a single distribution
}
\examples{
dat <- data.frame(meanlog = c(1, 1.5),
                  sdlog = c(1, 2),
                  weight = c(2, 1))
combine_lognorm(dat)
}
