% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{combine_norm}
\alias{combine_norm}
\title{Weight a set of normal parameters into a single distribution}
\usage{
combine_norm(dat)
}
\arguments{
\item{dat}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Given a set of arbitrary parameters that includes at least a weight column,
  take a weighted average of all the other parameters.
}
\examples{
dat <- data.frame(mean = c(10, 20, 30),
              sd = c(4, 5, 10),
              weight = c(2, 1, 2))
combine_norm(dat)
}
