% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{combine_scenario_parameters}
\alias{combine_scenario_parameters}
\title{Combine multiple SME distributions into a single unified view}
\usage{
combine_scenario_parameters(scenario_parameters)
}
\arguments{
\item{scenario_parameters}{Fitted scenario factors for individual SMEs.}
}
\value{
A dataframe.
}
\description{
Given a dataframe with multiple SME fitted distributions for a single
  scenario, decompose the lognormal distribution into normal parameters,
  apply weighting for opinion pooling, and construct a final combined
  distribution for each OpenFAIR scenario factor.
}
\examples{
NULL
}
