% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{derive_controls}
\alias{derive_controls}
\title{Generate the quantified capability parameters for a scenario}
\usage{
derive_controls(capability_ids, capability_parameters)
}
\arguments{
\item{capability_ids}{Comma-delimited list of capability ids}

\item{capability_parameters}{Dataframe of fitted and combined capability parameters}
}
\value{
A list.
}
\description{
Based on the \code{\link[evaluator]{derive_controls}} function
}
\details{
Creates the difficulty parameters (embedded list) for quantitative
  parameters.
}
\examples{
NULL
}
\seealso{
\code{\link[evaluator]{derive_controls}}
}
