% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_controls.R
\name{derive_controls_alt}
\alias{derive_controls_alt}
\title{Generate the quantified capability parameters for a scenario}
\usage{
derive_controls_alt(capability_ids, capabilities)
}
\arguments{
\item{capability_ids}{A dataframe.}

\item{capabilities}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Based on the \code{\link[evaluator]{derive_controls}} function
}
\details{
This custom derive_controls function allows for non-numeric capability IDs.

Creates the difficulty parameters (embedded dataframe) for quantitative
parameters.
}
\examples{
NULL
}
\seealso{
\code{\link[evaluator]{derive_controls}}
}
