% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enforce_questions}
\alias{enforce_questions}
\title{Validate that the parameter passed is a questions object}
\usage{
enforce_questions(x)
}
\arguments{
\item{x}{An object}
}
\description{
Validate that the parameter passed is a questions object
}
\examples{
\dontrun{
questions <- read_questions()
enforce_questions(questions)
}
}
