% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enforce_responses}
\alias{enforce_responses}
\title{Validate that the parameter passed is a responses object}
\usage{
enforce_responses(x)
}
\arguments{
\item{x}{An object}
}
\description{
Validate that the parameter passed is a responses object
}
\examples{
\dontrun{
answers <- read_answers()
enforce_answers(answers)
}
}
