% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enforce_tidyrisk_response_set}
\alias{enforce_tidyrisk_response_set}
\title{Validate that the parameter passed is a tidyrisk_response_set object}
\usage{
enforce_tidyrisk_response_set(x)
}
\arguments{
\item{x}{An object}
}
\description{
Validate that the parameter passed is a tidyrisk_response_set object
}
\examples{
\dontrun{
answers <- read_answers()
enforce_answers(answers)
}
}
