% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{fit_capabilities}
\alias{fit_capabilities}
\title{Fit SME capability estimates to distribution parameters}
\usage{
fit_capabilities(capability_answers)
}
\arguments{
\item{capability_answers}{Capability answers.}
}
\value{
A dataframe.
}
\description{
Fit SME capability estimates to distribution parameters
}
\examples{
NULL
}
