% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{fit_capabilities_geomean}
\alias{fit_capabilities_geomean}
\title{Fit capability parameters via a geometric mean}
\usage{
fit_capabilities_geomean(capabilities_answers)
}
\arguments{
\item{capabilities_answers}{Answers dataframe.}
}
\value{
A dataframe.
}
\description{
Fit capability parameters via a geometric mean
}
\examples{
NULL
}
