% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{fit_lognorm}
\alias{fit_lognorm}
\title{Find parameters that fit quantile values of an unknown lognormal distribution}
\usage{
fit_lognorm(low, high)
}
\arguments{
\item{low}{5th quantile.}

\item{high}{95th quantile.}
}
\value{
A dataframe.
}
\description{
With a 5th and 95th quantile point estimates, fit a lognormal distribution,
  returning the parameters of the distribution.
}
\examples{
NULL
}
