% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{fit_norm_trunc}
\alias{fit_norm_trunc}
\title{Find parameters that fit quantile values of an unknown truncated normal
  distribution}
\usage{
fit_norm_trunc(low, high, min = 0, max = Inf)
}
\arguments{
\item{low}{5th quantile.}

\item{high}{95th quantile.}

\item{min}{Lower bound of support.}

\item{max}{Upper bound of support.}
}
\value{
Dataframe.
}
\description{
With a 5th and 95th quantile point estimates and optional lower and
  upper bounds, fit a truncated normal distribution, returning the parameters of
  the distribution.
}
\examples{
fit_norm_trunc(low = 10, high = 50, min = 0, max = 100)
}
