% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{fit_scenarios_geomean}
\alias{fit_scenarios_geomean}
\title{Fit scenario parameters by applying a geometric mean}
\usage{
fit_scenarios_geomean(scenario_answers)
}
\arguments{
\item{scenario_answers}{Scenario answers.}
}
\value{
A dataframe.
}
\description{
Fit scenario parameters by applying a geometric mean
}
\examples{
NULL
}
