% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{fit_threat_communities}
\alias{fit_threat_communities}
\title{Fit each of the threat communities to a distribution}
\usage{
fit_threat_communities(threat_communities)
}
\arguments{
\item{threat_communities}{Dataframe of threat communities.}
}
\value{
A dataframe.
}
\description{
Fit each of the threat communities to a distribution
}
\examples{
NULL
}
