% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{generate_cost_function}
\alias{generate_cost_function}
\title{Generate a sum of squares cost function for optimization}
\usage{
generate_cost_function(func)
}
\arguments{
\item{func}{A distribution function.}
}
\value{
A function.
}
\description{
This is an internal helper function that generates a sum of squares
  cost function for any given `r*` function (e.g. rnorm, rlognorm). The
  resulting function is intended to be used by an `optim` call for fitting
  quantiles to distribution parameters.
}
\examples{
NULL
}
