% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weights.R
\name{generate_weights}
\alias{generate_weights}
\title{Generate a weighting table for SMEs based upon their calibration answers}
\usage{
generate_weights(questions, responses)
}
\arguments{
\item{questions}{tidyrisk_question_set object.}

\item{responses}{tidyrisk_response_set object}
}
\value{
A dataframe of SMEs and their numerical weighting.
}
\description{
Generate a weighting table for SMEs based upon their calibration answers
}
\examples{
NULL
}
