% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_smes_domains}
\alias{get_smes_domains}
\title{Calculate the prioritized list of domains for a given SME}
\usage{
get_smes_domains(sme, domains, expertise)
}
\arguments{
\item{sme}{Name of SME.}

\item{domains}{Domains dataframe.}

\item{expertise}{Expertise dataframe.}
}
\value{
A vector.
}
\description{
With a dataframe of all domains, a datafrme of the prioritized list of
  domains to focus on for a SME, and the name of a specific SME of interest,
  return a vector of the domains in order of priority.
}
\examples{
NULL
}
