% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{is_questions}
\alias{is_questions}
\title{Test if the object is a questions}
\usage{
is_questions(x)
}
\arguments{
\item{x}{An object}
}
\description{
This function returns TRUE for questions or subclasses therof, and FALSE for all other objects.
}
\examples{
\dontrun{
is_questions(x)
}
}
