% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{is_responses}
\alias{is_responses}
\title{Test if the object is a responses}
\usage{
is_responses(x)
}
\arguments{
\item{x}{An object}
}
\description{
This function returns TRUE for responses or subclasses therof, and FALSE for all other objects.
}
\examples{
\dontrun{
is_responses(x)
}
}
