% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_response_set.R
\name{is_tidyrisk_response_set}
\alias{is_tidyrisk_response_set}
\title{Test if the object is a tidyrisk_response_set}
\usage{
is_tidyrisk_response_set(x)
}
\arguments{
\item{x}{An object}
}
\description{
This function returns TRUE for tidyrisk_response_set or sub-classes
  thereof, and FALSE for all other objects.
}
\examples{
\dontrun{
is_tidyrisk_response_set(x)
}
}
