% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{lognormal_to_normal}
\alias{lognormal_to_normal}
\title{Convert lognormal parameters to normal parameters}
\usage{
lognormal_to_normal(meanlog, sdlog)
}
\arguments{
\item{meanlog}{Mean log.}

\item{sdlog}{Standard deviation log.}
}
\value{
A list.
}
\description{
Given a set of parameters describing a lognormal distribution, return
  the parameters of the underlying normal distribution.
}
\examples{
lognormal_to_normal(meanlog=1, sdlog=3)
}
