% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_handouts.R
\name{make_handouts}
\alias{make_handouts}
\title{Create a set of interview handouts for a SME}
\usage{
make_handouts(sme, calibration_questions, domains, expertise, scenarios,
  capabilities, output_dir = getwd())
}
\arguments{
\item{sme}{Name of SME.}

\item{calibration_questions}{calibration questions dataframe.}

\item{domains}{domains dataframe.}

\item{expertise}{expertise dataframe.}

\item{scenarios}{scenarios dataframe.}

\item{capabilities}{expertise dataframe.}

\item{output_dir}{Directory to place output.}
}
\value{
Output of render.
}
\description{
Creates two MS Word formatted documents. One is an `answers` document
  that contains the answers to the calibration questions, the other does not
  contain answers and is intended to be a visual reference (and possible
  take away) for the SME.
}
\examples{
NULL
}
