% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_slides.R
\name{make_slides}
\alias{make_slides}
\title{Create interview slides}
\usage{
make_slides(sme, source_dir, output_dir = getwd(),
  assessment_title = "Strategic Risk Assessment")
}
\arguments{
\item{sme}{SME.}

\item{source_dir}{Directory location for input files.}

\item{output_dir}{Directory location for knitted slides.}

\item{assessment_title}{Title of the assessment being performed}
}
\value{
Output of render.
}
\description{
Creates a HTML-format slideshow for conducting and interview.
}
\examples{
NULL
}
