% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_calibration_answers}
\alias{mc_calibration_answers}
\title{MetroCare Hospital Calibration Answers}
\format{
A data frame with 50 rows and 5 variables:
\describe{
\item{sme}{name of the subject matter expert}
\item{calibration_id}{unique identifier of the calibration question}
\item{low}{SME's low end estimate}
\item{high}{SME's high end estimate}
\item{date}{date of answer}
}
}
\source{
This is hypothetical information. Any similarity to any other
entity is completely coincidental.
}
\usage{
mc_calibration_answers
}
\description{
A dataset of SME answers to calibration questions.
}
\keyword{datasets}
