% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_capabilities}
\alias{mc_capabilities}
\title{MetroCare Hospital Capabilities}
\format{A data frame with 60 rows and 3 variables:
\describe{
  \item{capability_id}{unique identifier of the capability}
  \item{domain_id}{domain associated with the capability}
  \item{capability}{text description of the capability}
}}
\source{
This is hypothetical information. Any similarity to any other
entity is completely coincidental.
}
\usage{
mc_capabilities
}
\description{
A dataset of program capabilities.
}
\keyword{datasets}
