% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_capability_parameters_fitted}
\alias{mc_capability_parameters_fitted}
\title{MetroCare Hospital Capability Parameters (fitted)}
\format{A data frame with 300 rows and 10 variables:
\describe{
\item{sme}{name of the sme providing the response}
\item{capability_id}{unique identifier}
\item{date}{text description of the threat community}
\item{capability_func}{capability sampling function}
\item{capability_mean}{capability mean}
\item{capability_sd}{capability standard deviation}
\item{capability_min}{capability minimum}
\item{capability_max}{capability maximum}
\item{low}{threat communities capability, high end}
\item{high}{threat communities capability, high end}
}}
\source{
This is hypothetical information. Any similarity to any other
entity is completely coincidental.
}
\usage{
mc_capability_parameters_fitted
}
\description{
A dataset of sample fitted capability parameters.
}
\keyword{datasets}
