% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_domains}
\alias{mc_domains}
\title{MetroCare Hospital Domains}
\format{A data frame with 14 rows and 2 variables:
\describe{
  \item{domain_id}{unique domain id}
  \item{domain}{domain title}
}}
\source{
This is hypothetical information. Any similarity to any other
entitiy is completely coincidental.
}
\usage{
mc_domains
}
\description{
A dataset of program domains.
}
\keyword{datasets}
