% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_scenario_answers}
\alias{mc_scenario_answers}
\title{MetroCare Hospital Scenario Answers}
\format{A data frame with 1 rows and 7 variables:
\describe{
  \item{sme}{name of the SME}
  \item{scenario_id}{identifier of the scenario}
  \item{freq_low}{frequency estimate, low}
  \item{freq_high}{frequency estimate, high}
  \item{imp_low}{impact estimate, low}
  \item{imp_high}{impact estimate, high}
  \item{date}{date of the answer}
}}
\source{
This is hypothetical information. Any similarity to any other
entity is completely coincidental.
}
\usage{
mc_scenario_answers
}
\description{
A dataset of SME answers to scenarios.
}
\keyword{datasets}
