% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_scenario_parameters_fitted}
\alias{mc_scenario_parameters_fitted}
\title{MetroCare Hospital Scenario Parameters (fitted)}
\format{A data frame with 280 rows and 17 variables:
\describe{
  \item{sme}{text title of the threat community}
  \item{scenario_id}{unique identifier}
  \item{date}{text description of the threat community}
  \item{impact_func}{threat communities capability, low end}
  \item{impact_meanlog}{threat communities capability, high end}
  \item{impact_sdlog}{type of the threat community}
  \item{impact_min}{action type of the threat community}
  \item{impact_max}{action type of the threat community}
  \item{imp_low}{action type of the threat community}
  \item{imp_high}{action type of the threat community}
  \item{frequency_func}{threat communities capability, low end}
  \item{frequency_meanlog}{threat communities capability, high end}
  \item{frequency_sdlog}{type of the threat community}
  \item{frequency_min}{action type of the threat community}
  \item{frequency_max}{action type of the threat community}
  \item{freq_low}{action type of the threat community}
  \item{freq_high}{action type of the threat community}
}}
\source{
This is hypothetical information. Any similarity to any other
entity is completely coincidental.
}
\usage{
mc_scenario_parameters_fitted
}
\description{
A dataset of sample fitted scenario parameters.
}
\keyword{datasets}
