% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_scenarios}
\alias{mc_scenarios}
\title{MetroCare Risk Scenarios}
\format{A data frame with 56 rows and 5 variables:
\describe{
  \item{scenario_id}{unique identifier}
  \item{scenario}{scenario description}
  \item{tcomm}{name of threat community}
  \item{domain_id}{domain id}
  \item{controls}{comma seperated list of control ids}
}}
\source{
This is hypothetical information. Any similarity to any other
entitiy is completely coincidental.
}
\usage{
mc_scenarios
}
\description{
A dataset of sample risk scenarios.
}
\keyword{datasets}
