% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_threat_parameters_fitted}
\alias{mc_threat_parameters_fitted}
\title{MetroCare Hospital Threat Parameters (fitted)}
\format{A data frame with 8 rows and 12 variables:
\describe{
  \item{action_type}{action type}
  \item{category}{category}
  \item{definition}{text description of the threat community}
  \item{high}{action type of the threat community}
  \item{low}{type of the threat community}
  \item{threat_community}{text title of the threat community}
  \item{threat_func}{sampling function}
  \item{threat_id}{unique identifier}
  \item{threat_max}{sampling function}
  \item{threat_mean}{sampling function}
  \item{threat_sd}{sampling function}
  \item{threat_min}{sampling function}
}}
\source{
This is hypothetical information. Any similarity to any other
entity is completely coincidental.
}
\usage{
mc_threat_parameters_fitted
}
\description{
A dataset of sample fitted threat parameters.
}
\keyword{datasets}
