% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{normal_to_lognormal}
\alias{normal_to_lognormal}
\title{Convert normal paramaters to lognormal parameters}
\usage{
normal_to_lognormal(normmean, normsd)
}
\arguments{
\item{normmean}{Mean.}

\item{normsd}{Standard deviation.}
}
\value{
A list.
}
\description{
Given parameters that describe a normal distribution, convert them back
  to parameters for a lognormal distribution.
}
\examples{
normal_to_lognormal(normmean = 20, normsd = 3)
}
