% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{questions}
\alias{questions}
\alias{new_questions}
\alias{as.questions}
\title{Construct a questions object}
\usage{
questions(domains, scenarios, capabilities, calibration, expertise,
  threat_communities)

new_questions(x)

as.questions(x, ...)
}
\arguments{
\item{domains}{Domains}

\item{scenarios}{Scenario questions}

\item{capabilities}{Capability questions}

\item{calibration}{Calibration questions}

\item{expertise}{SME expertise}

\item{threat_communities}{Threat communities}

\item{x}{object to coerce}

\item{...}{Individual dataframes}
}
\description{
\code{new.questions} is a low-level constructor that takes a list of dataframes.
\code{questions} constructs a questions object from dataframes.
\code{as.blob} is a S3 generic that converts existing objects.
}
\examples{
NULL
}
