% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_answers}
\alias{read_answers}
\title{Read all SMEs answers}
\usage{
read_answers(source_dir = getwd())
}
\arguments{
\item{source_dir}{Directory location where input files are found.}
}
\value{
A responses object
}
\description{
Reads in all the answers recorded to the calibration, scenarios, and
  capability questions.
}
\examples{
\dontrun{
read_answers()
}
}
