% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_data}
\alias{read_data}
\title{Read data files}
\usage{
read_data(source_dir = getwd(), active_only = TRUE)
}
\arguments{
\item{source_dir}{Directory location to find input files.}

\item{active_only}{Read in only the active elements, defaults to TRUE.}
}
\value{
A list.
}
\description{
Read data files
}
\examples{
NULL
}
