% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_questions}
\alias{read_questions}
\title{Read scenario questions}
\usage{
read_questions(source_dir = getwd(), active_only = TRUE)
}
\arguments{
\item{source_dir}{Directory location to find input files.}

\item{active_only}{Read in only the active elements, defaults to TRUE.}
}
\value{
A list.
}
\description{
Read scenario questions
}
\examples{
NULL
}
