% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_responses}
\alias{read_responses}
\title{Read all SMEs responses}
\usage{
read_responses(source_dir = getwd())
}
\arguments{
\item{source_dir}{Directory location where input files are found.}
}
\value{
A tidyrisk_response_set object
}
\description{
Reads in all the responses recorded to the calibration, scenarios, and
  capability questions.
}
\examples{
\dontrun{
read_responses()
}
}
