% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{responses}
\alias{responses}
\alias{new_responses}
\alias{as.responses}
\title{Construct a responses object}
\usage{
responses(...)

new_responses(calibration_answers, scenario_answers, capability_answers)

as.responses(x, ...)
}
\arguments{
\item{...}{Individual dataframes}

\item{calibration_answers}{Calibration responses}

\item{scenario_answers}{Scenarios responses}

\item{capability_answers}{Capability responses}

\item{x}{object to coerce}
}
\description{
\code{new.responses} is a low-level constructor that takes a list of dataframes.
\code{responses} constructs a responses from dataframes.
\code{as.blob} is a S3 generic that converts existing objects.
}
\examples{
NULL
}
