% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collector.R
\name{set_collector}
\alias{set_collector}
\title{Set functions to use collector}
\usage{
set_collector(funs = NULL, pkg = NULL, path = "collector")
}
\arguments{
\item{funs}{Names of functions, all exported functions are considered by default}

\item{pkg}{Package names, can be omitted if called from .onLoad() or if we
can file a DESCRIPTION file in the working directory}

\item{path}{The path to a directory where to save the collected data

For each call A file will be saved, it can be read with \code{qs::qread()} and
contains a list with 2 elements named \code{call} and \code{env}.}
}
\description{
After being setup, calls to these functions will trigger a serialization
of the calls and necessary environmments.
}
