% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_question_set.R
\name{tidyrisk_question_set}
\alias{tidyrisk_question_set}
\alias{new_tidyrisk_question_set}
\alias{as.tidyrisk_question_set}
\alias{validate_tidyrisk_question_set}
\title{Construct a tidyrisk_question_set object}
\usage{
tidyrisk_question_set(domains, scenarios, capabilities, calibration,
  expertise, threat_communities)

new_tidyrisk_question_set(x)

as.tidyrisk_question_set(x, ...)

validate_tidyrisk_question_set(x)
}
\arguments{
\item{domains}{Domains}

\item{scenarios}{Scenario questions}

\item{capabilities}{Capability questions}

\item{calibration}{Calibration questions}

\item{expertise}{SME expertise}

\item{threat_communities}{Threat communities}

\item{x}{object to coerce}

\item{...}{Individual dataframes}
}
\description{
\code{new.tidyrisk_question_set} is a low-level constructor that takes a list of dataframes.
\code{tidyrisk_question_set} constructs a tidyrisk_question_set object from dataframes.
\code{as.tidyrisk_question_set} is a S3 generic that converts existing objects.
\code{validate_tidyrisk_question_set} verifies that the data elements are internally consistent.
}
\examples{
NULL
}
