% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_response_set.R
\name{tidyrisk_response_set}
\alias{tidyrisk_response_set}
\alias{new_tidyrisk_response_set}
\alias{as.tidyrisk_response_set}
\title{Construct a tidyrisk_response_set object}
\usage{
tidyrisk_response_set(
  calibration_answers,
  scenario_answers,
  capability_answers
)

new_tidyrisk_response_set(
  calibration_answers,
  scenario_answers,
  capability_answers
)

as.tidyrisk_response_set(x, ...)
}
\arguments{
\item{calibration_answers}{Calibration tidyrisk_response_set}

\item{scenario_answers}{Scenarios tidyrisk_response_set}

\item{capability_answers}{Capability tidyrisk_response_set}

\item{x}{object to coerce}

\item{...}{Individual dataframes}
}
\description{
\code{new.tidyrisk_response_set} is a low-level constructor that takes a list of dataframes.
\code{tidyrisk_response_set} constructs a tidyrisk_response_set from dataframes.
\code{as.tidyrisk_response_set} is a S3 generic that converts existing objects.
}
\examples{
NULL
}
