% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate a question set}
\usage{
validate(questions)
}
\arguments{
\item{questions}{a tidyrisk_question_set object}
}
\value{
Invisibly returns TRUE if all checks pass or FALSE if any checks fail
}
\description{
Confirms that all scenario data is well formed (internally consistent).
  This is a good sanity check after reading in new scenario data to ensure
  that no data entry errors occurred.
}
\examples{
\dontrun{
validate(questions)
}
}
