% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_readability}
\alias{check_readability}
\title{Check the readability of scenario text}
\usage{
check_readability(x)
}
\arguments{
\item{x}{Scenarios.}
}
\value{
A dataframe of the scenario id, domain, and the Flesch-Kincaid readability score.
}
\description{
With a dataframe of scenario text, scenario_id, and domain, generate the Flesch-
  Kincaid score and return that score along with the scenario ID and domain
  as a tidy dataframe.
}
\examples{
\dontrun{
questions <- read_questions()
check_readability(questions)
}
}
