% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collector.R
\docType{package}
\name{collector}
\alias{collector}
\alias{collector-package}
\title{\code{collector} package}
\description{
Quantified Risk Assessment Data Collection
}
\details{
See the README on
\href{https://cran.r-project.org/package=collector/README.html}{CRAN}
or \href{https://github.com/davidski/collector}{GitHub}
}
