% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{fit_capabilities_geomean}
\alias{fit_capabilities_geomean}
\title{Fit capability parameters via a geometric mean}
\usage{
fit_capabilities_geomean(capabilities_answers)
}
\arguments{
\item{capabilities_answers}{Answers dataframe.}
}
\value{
A dataframe.
}
\description{
Fit capability parameters via a geometric mean
}
\examples{
data(mc_capability_answers)
fit_capabilities_geomean(mc_capability_answers)
}
\seealso{
Other distribution fitting functions: \code{\link{combine_lognorm_trunc}},
  \code{\link{combine_lognorm}},
  \code{\link{combine_norm}},
  \code{\link{fit_capabilities}},
  \code{\link{fit_lognorm_trunc}},
  \code{\link{fit_lognorm}}, \code{\link{fit_norm_trunc}},
  \code{\link{fit_pois}},
  \code{\link{fit_scenarios_geomean}},
  \code{\link{fit_scenarios}},
  \code{\link{fit_threat_communities}},
  \code{\link{generate_cost_function}},
  \code{\link{lognormal_to_normal}},
  \code{\link{normal_to_lognormal}}
}
\concept{distribution fitting functions}
