% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{fit_pois}
\alias{fit_pois}
\title{Find parameters that fit a poisson distribution.}
\usage{
fit_pois(low, high)
}
\arguments{
\item{low}{5th quantile.}

\item{high}{95th quantile.}
}
\value{
A dataframe.
}
\description{
With a 5th and 95th quantile point estimates and optional lower and
  upper bounds, fit a poisson distribution, returning the parameters of
  the distribution.
}
\examples{
fit_pois(low = 10, high = 50)
}
