% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{fit_scenarios_geomean}
\alias{fit_scenarios_geomean}
\title{Fit scenario parameters by applying a geometric mean}
\usage{
fit_scenarios_geomean(scenario_answers)
}
\arguments{
\item{scenario_answers}{Scenario answers.}
}
\value{
A dataframe.
}
\description{
Fit scenario parameters by applying a geometric mean
}
\examples{
NULL
}
\seealso{
Other distribution fitting functions: \code{\link{combine_lognorm_trunc}},
  \code{\link{combine_lognorm}},
  \code{\link{combine_norm}},
  \code{\link{fit_capabilities_geomean}},
  \code{\link{fit_capabilities}},
  \code{\link{fit_lognorm_trunc}},
  \code{\link{fit_lognorm}}, \code{\link{fit_norm_trunc}},
  \code{\link{fit_pois}}, \code{\link{fit_scenarios}},
  \code{\link{fit_threat_communities}},
  \code{\link{generate_cost_function}},
  \code{\link{lognormal_to_normal}},
  \code{\link{normal_to_lognormal}}
}
\concept{distribution fitting functions}
