% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_handouts.R
\name{make_handouts}
\alias{make_handouts}
\title{Create a set of interview handouts for a SME}
\usage{
make_handouts(sme, questions, output_dir = getwd())
}
\arguments{
\item{sme}{Name of the SME.}

\item{questions}{questions object}

\item{output_dir}{Directory to place output.}
}
\value{
Output of render.
}
\description{
Creates two MS Word formatted documents. One is an `answers` document
  that contains the answers to the calibration questions, the other does not
  contain answers and is intended to be a visual reference (and possible
  take away) for the SME.
}
\examples{
\dontrun{
questions <- read_questions()
make_handouts("Sally Expert", questions)
}
}
