% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_sme_top_domains}
\alias{mc_sme_top_domains}
\title{MetroCare Hospital SME Top Domains}
\format{A data frame with 5 rows and 7 variables:
\describe{
  \item{sme}{sme name}
  \item{domain1}{sme's first domain}
  \item{domain2}{sme's second domain}
  \item{domain3}{sme's third domain}
  \item{domain4}{sme's fourth domain}
  \item{domain5}{sme's fifth domain}
  \item{domain6}{sme's sixth domain}
}}
\source{
This is hypothetical information. Any similarity to any other
entitiy is completely coincidental.
}
\usage{
mc_sme_top_domains
}
\description{
A dataset of focus domains per SME.
}
\keyword{datasets}
