% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distributions.R
\name{normal_to_lognormal}
\alias{normal_to_lognormal}
\title{Convert normal parameters to lognormal parameters}
\usage{
normal_to_lognormal(normmean, normsd)
}
\arguments{
\item{normmean}{Mean.}

\item{normsd}{Standard deviation.}
}
\value{
A list.
}
\description{
Given parameters that describe a normal distribution, convert them back
  to parameters for a lognormal distribution.
}
\examples{
normal_to_lognormal(normmean = 20, normsd = 3)
}
\seealso{
Other distribution fitting functions: \code{\link{combine_lognorm_trunc}},
  \code{\link{combine_lognorm}},
  \code{\link{combine_norm}},
  \code{\link{fit_capabilities_geomean}},
  \code{\link{fit_capabilities}},
  \code{\link{fit_lognorm_trunc}},
  \code{\link{fit_lognorm}}, \code{\link{fit_norm_trunc}},
  \code{\link{fit_pois}},
  \code{\link{fit_scenarios_geomean}},
  \code{\link{fit_scenarios}},
  \code{\link{fit_threat_communities}},
  \code{\link{generate_cost_function}},
  \code{\link{lognormal_to_normal}}
}
\concept{distribution fitting functions}
