% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Create a quantitative_scenarios dataframe suitable for `evaluator`}
\usage{
prepare_data(scenario_parameters, capability_parameters, threat_parameters,
  questions)
}
\arguments{
\item{scenario_parameters}{Scenarios with final parameters defined.}

\item{capability_parameters}{Capabilities with final parameters defined.}

\item{threat_parameters}{Threat communities with final parameters defined.}

\item{questions}{A questions object.}
}
\value{
A dataframe.
}
\description{
Create a quantitative_scenarios dataframe suitable for `evaluator`
}
\examples{
\dontrun{
questions <- read_questions()
quantitative_scenarios <- prepare_data(scenario_parameters,
      capability_parameters, threat_parameters, questions)
}
}
