% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Create one or more quantitative scenarios objects suitable for simulation by `evaluator`}
\usage{
prepare_data(scenario_parameters, capability_parameters, threat_parameters,
  questions)
}
\arguments{
\item{scenario_parameters}{Scenarios with final parameters defined.}

\item{capability_parameters}{Capabilities with final parameters defined.}

\item{threat_parameters}{Threat communities with final parameters defined.}

\item{questions}{A questions object.}
}
\value{
A list of \code{tidyrisk_scenario} objects.
}
\description{
Given parameters for the scenarios, threat communities, capabilities, and
the question set, generate a list of `tidyrisk_scenario` objects that may be
fed into \code{evaluator::run_simulations()} for Monte Carlo simulation.
}
\examples{
\dontrun{
question_set <- questions(mc_domains, mc_capabilities, mc_scenarios,
                          mc_sme_top_domains, calibration_questions,
                          mc_threat_communities))
      capability_parameters, threat_parameters, questions)
}
}
