% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_questions}
\alias{read_questions}
\title{Read scenario questions}
\usage{
read_questions(source_dir = getwd(), active_only = TRUE)
}
\arguments{
\item{source_dir}{Directory location to find input files.}

\item{active_only}{Read in only the active elements, defaults to TRUE.}
}
\value{
A tidyrisk_question_set object
}
\description{
Reads in all the questions for which subject matter expert input is
  needed. Includes the domains, capabilities, scenarios, calibration
  questions, and threat communities.
}
\examples{
\dontrun{
read_questions()
}
}
